Fix message details (chat)
