+++ src/app/app-routing.module.ts
@@ -8,6 +8,7 @@ import {BoardUserComponent} from './board-user/board-user.component';
 import {BoardAdminComponent} from './board-admin/board-admin.component';
 import {PageNotFoundComponent} from "./page-not-found/page-not-found.component";
 import {ContactModalComponent} from "./contact-modal/contact-modal.component";
+import {AdChatComponent} from "./ad-chat/ad-chat.component";
 import {AdMessagesDetailsComponent} from "./ad-messages-details/ad-messages-details.component";
 
 const routes: Routes = [
@@ -19,7 +20,8 @@ const routes: Routes = [
   {path: 'contactModal/:username', component: ContactModalComponent},
   {path: '**', component: PageNotFoundComponent},
   {path: '', redirectTo: 'home', pathMatch: 'full'},
-  {path: 'adMessagesDetails', component: AdMessagesDetailsComponent},
+  {path: 'adChat', component: AdChatComponent},
+  {path: 'adChat/adMessagesDetails/:adId', component: AdMessagesDetailsComponent},
 ];
 
 @NgModule({
