+++ src/app/app.component.html
@@ -38,7 +38,7 @@
         </a>
         <ul class="dropdown-menu dropdown-menu-dark text-small shadow show" aria-labelledby="dropdownUser" style=""
             data-popper-placement="top-start">
-          <li><a href="/adMessages" class="dropdown-item">Messages</a></li>
+          <li><a href="/adChat" class="dropdown-item">Messages</a></li>
           <li><a href="/profile" class="dropdown-item">Profile</a></li>
           <li>
             <hr class="dropdown-divider">
