+++ src/app/messageDataService.ts
@@ -1,4 +1,5 @@
 import { Message } from './message';
 export class MessageDataService {
   public messages: Message[];
+  public adId: number;
 }
